backend-python/requirements.txt
azure-storage-blob==12.19.1
python-dotenv==1.0.1
mysql-connector-python==8.4.0
transformers==4.41.2
torch==2.3.0
Pillow==10.3.0
requests==2.31.0
